Refactor agent activity renderers into modules (#609)
Split tool-renderers.tsx and message-renderers.tsx into smaller modules
to reduce complexity and improve maintainability.
